+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 # [Released]
 
+# [v1.1.8]
+
+### Added
+
+### Changed
+
+### Fixed
+
+## Removed
+
 # [v1.1.7]
 
 ### Added